// src/app/core/models/clinical-notes/consult.model.ts
export interface ConsultObjectiveToothTesting {
  number: string;
  cold?: string;
  heat?: string;
  ept?: string;
  perc?: string;
  palp?: string;
  mob?: string;
  prob?: string;
  bite?: string;
  restoration?: string;
  other?: string;
}

export interface ConsultDiagnosis extends ConsultObjectiveToothTesting { }

export interface ConsultTreatmentPlan extends ConsultObjectiveToothTesting { }

export interface CreateConsultDto {
  txProvider: string;
  txAssistant: string;
  refferingDentist: string;
  noSymptoms?: boolean;
  sensitivityToCold?: boolean;
  sensitivityToHeat?: boolean;
  swelling?: boolean;
  sensitivityToPressure?: boolean;
  rctFinished?: boolean;
  locationUr?: boolean;
  locationUl?: boolean;
  locationLr?: boolean;
  locationLL?: boolean;
  chiefNotes?: string;
  bp?: string;
  pulse?: string;
  temp?: string;
  allergies?: string;
  details?: string;
  dentalCare?: string;
  dentalPhobia?: string;
  recentDentalWork?: string;
  plannedDentalWork?: string;
  pastPainIntensity?: string;
  painSharp?: boolean;
  painDull?: boolean;
  painAchy?: boolean;
  painThrobbing?: boolean;
  painStabbing?: boolean;
  painElectric?: boolean;
  painPressure?: boolean;
  presentPainIntensity?: string;
  painIntermittent?: boolean;
  painConstant?: boolean;
  painSpontaneous?: boolean;
  painProvoked?: boolean;
  painDiurnal?: boolean;
  painNocturnal?: boolean;
  painDuration?: string;
  painDurationOtcHelp?: string;
  painDurationRxHelp?: string;
  painDurationAbxHelp?: string;
  painDurationIceWaterHelp?: string;
  hxOfSwelling?: string;
  hxOfBruxing?: string;
  rxAnalgesicsHelp?: string;
  abxHelp?: string;
  iceWaterHelp?: string;
  subjectiveNotes?: string;
  extraOralExam?: boolean;
  extraOralSwellingL?: boolean;
  extraOralSwellingR?: boolean;
  extraOralSwellingTxt?: string;
  extraOralTMJExamL?: boolean;
  extraOralTMJExamR?: boolean;
  extraOralTMJExamTxt?: string;
  extraOralMsOfMasticationsL?: boolean;
  extraOralMsOfMasticationsR?: boolean;
  extraOralMsOfMasticationsTxt?: string;
  intraOralExam?: boolean;
  intraOralSinus?: string;
  intraOralMissingTeeth?: string;
  intraOralCaries?: string;
  intraOralSwelling?: string;
  intraOralSwellingLocation?: string;
  intraOralCracks?: string;
  objectiveToothTesting?: ConsultObjectiveToothTesting[];
  objectiveNotes?: string;
  diagnosisList?: ConsultDiagnosis[];
  treatmentPlanList?: ConsultTreatmentPlan[];
  diagnosisOtherNote?: string;
  pulpNormal?: boolean;
  pulpNormalTxt?: string;
  pulpStones?: boolean;
  pulpStonesTxt?: string;
  pulpPartialCalcification?: boolean;
  pulpPartialCalcificationTxt?: string;
  pulpCompleteCalcification?: boolean;
  pulpCompleteCalcificationTxt?: string;
  pulpInternalResorption?: boolean;
  pulpInternalResorptionTxt?: string;
  pulpExternalResorption?: boolean;
  pulpExternalResorptionTxt?: string;
  pulpOpenApex?: boolean;
  pulpOpenApexTxt?: string;
  pulpExcessiveRootCurvature?: boolean;
  pulpExcessiveRootCurvatureTxt?: string;
  previousGp?: boolean;
  previousGpTxt?: string;
  previousCarriers?: boolean;
  previousCarriersTxt?: string;
  previousPaste?: boolean;
  previousPasteTxt?: string;
  previousAg?: boolean;
  previousAgTxt?: string;
  previousPerforation?: boolean;
  previousPerforationTxt?: string;
  previousUnderfield?: boolean;
  previousUnderfieldTxt?: string;
  previousOverfilled?: boolean;
  previousOverfilledTxt?: string;
  previousSeparatedInstrument?: boolean;
  previousSeparatedInstrumentTxt?: string;
  previousUncleanedCanals?: boolean;
  previousUncleanedCanalsTxt?: string;
  otherGp?: boolean;
  otherGpTxt?: string;
  otherCarriers?: boolean;
  otherCarriersTxt?: string;
  otherPaste?: boolean;
  otherPasteTxt?: string;
  otherAg?: boolean;
  otherAgTxt?: string;
  otherPerforation?: boolean;
  otherPerforationTxt?: string;
  horizontalBoneLoss?: string;
  verticalBoneloss?: string;
  pariapicalNormal?: boolean;
  pariapicalNormalTxt?: string;
  pariapicalWidenedPdl?: boolean;
  pariapicalWidenedPdlTxt?: string;
  pariapicalApicalRl?: boolean;
  pariapicalApicalRlTxt?: string;
  pariapicalLateralRl?: boolean;
  pariapicalLateralRlTxt?: string;
  pariapicalCondensing?: boolean;
  pariapicalCondensingTxt?: string;
  caries?: boolean;
  radiographicNotes?: string;
}